Update non-scalar data handling.
Ensemble behavior is implicit when gmxapi input/output data references
(including parameters) are set to Python iterables. There is an
immediate data typing need, then, to explicitly describe non-scalar data
with references to N-Dimensional Arrays (gmx.NDArray).

This commit updates the syntax and examples to use Python iterables
where AllScatter or Map type behavior is intended and to use a
numpy-like NDArray constructor for parameters that are passed as arrays
at the point of execution.