Change `extract()` to `result()`.
`result()` is more consistent with "futures" sorts of paradigms to
indicate that a proxy to computation is converted to the result of the
computation, whereas `extract()` is more likely to refer to type-casting
and/or in-memory transformations.

Note: there is a slight distinction here between, say,
`concurrent.futures.Future` and gmxapi providers of `result()` in that
our `result()` method is available on handles to tasks even before they
have been `submit`ed, and can trigger implicit submission. We might
think of this as a distinction between (typical) "eager" task launch
versus (gmxapi) "lazy" task launch, which doesn't seem sufficiently
relevant to affect this user interface decision.